chore: move flatten, remap_keys and match to tools
