+++ multikeydict/tools/__init__.py
@@ -1,1 +1,4 @@
+from .flatten import flatten
+from .map import remap_items
+from .match import match_keys
 from .reorder_key import reorder_key
